* Updates to weapon sound and weapon scripts
Changes to weapon scripts that define particle names, updated to reflect the recent cleanup.
Changes to weapon sounds, specifically rocket trail and conc grenade sounds, rocket  trail is now properly able to fade with distance and the conc grenade has a lower volume and random pitch to help when there are multiple concs in the world. Distance fading seems to not work as the sound is emitted from the player/weapon and not the grenade, but the sound is ready for the possible change